app/driver/requirements.txt
pyyaml
hidapi==0.15.0
tabulate==0.9.0